Remove unnecessary component layer for task overview side panel
